+++ CHANGELOG.rst
@@ -1,6 +1,6 @@
 3.23.0
 ======
-Unreleased
+April 6, 2020
 
 Features
 --------